WIP: replace old vpn client
