+++ nym-vpn-core/crates/nym-vpn-api-client/src/lib.rs
@@ -1,17 +1,18 @@
-pub mod account;
 pub mod responses;
 
 mod client;
 mod error;
 pub(crate) mod headers;
-mod helpers;
+// mod helpers;
 pub(crate) mod jwt;
+mod request;
 mod routes;
+mod types;
 
-pub use client::{Client, ClientBuilder, VpnApiClientExt, VpnApiError};
+pub use client::{AccountClient};
 pub use error::VpnApiClientError;
-pub use helpers::{
-    get_countries, get_entry_countries, get_entry_gateways, get_exit_countries, get_exit_gateways,
-    get_gateways,
-};
-pub use responses::{Country, Gateway, Location};
+// pub use helpers::{
+//     get_countries, get_entry_countries, get_entry_gateways, get_exit_countries, get_exit_gateways,
+//     get_gateways,
+// };
+pub use types::{Country, Gateway, Location};
