+++ nym-vpn-core/crates/nym-vpn-api-client/src/request.rs
@@ -1,0 +1,24 @@
+// Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct RegisterDeviceRequestBody {
+    pub device_identity_key: String,
+    pub signature: String,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct RequestZkNymRequestBody {
+    pub blinded_signing_request_base58: String,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct CreateSubscriptionRequestBody {
+    pub valid_from_utc: String,
+    pub subscription_kind: String,
+}
