+++ nym-vpn-core/crates/nym-vpn-api-client/src/routes.rs
@@ -1,5 +1,16 @@
+// Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
+pub const PUBLIC: &str = "public";
+pub const V1: &str = "v1";
+pub const ACCOUNT: &str = "account";
+pub const SUMMARY: &str = "summary";
+pub const DEVICE: &str = "device";
+pub const ACTIVE: &str = "active";
+pub const ZKNYM: &str = "zknym";
+pub const SUBSCRIPTION: &str = "subscription";
 pub const DIRECTORY: &str = "directory";
 pub const GATEWAYS: &str = "gateways";
+pub const COUNTRIES: &str = "countries";
 pub const ENTRY: &str = "entry";
 pub const EXIT: &str = "exit";
-pub const COUNTRIES: &str = "countries";
