+++ nym-vpn-core/crates/nym-vpn-api-client/src/types.rs
@@ -1,0 +1,188 @@
+// Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
+use std::sync::Arc;
+
+use nym_crypto::asymmetric::ed25519;
+use nym_validator_client::{signing::signer::OfflineSigner as _, DirectSecp256k1HdWallet};
+use serde::{Deserialize, Serialize};
+
+use crate::jwt::Jwt;
+
+const MAX_PROBE_RESULT_AGE_MINUTES: i64 = 60;
+
+pub struct Account {
+    wallet: DirectSecp256k1HdWallet,
+}
+
+impl Account {
+    #[allow(unused)]
+    fn random() -> Self {
+        let mnemonic = bip39::Mnemonic::generate(24).unwrap();
+        let wallet = DirectSecp256k1HdWallet::from_mnemonic("n", mnemonic);
+        Self { wallet }
+    }
+
+    pub fn id(&self) -> String {
+        self.wallet.get_accounts().unwrap()[0].address().to_string()
+    }
+
+    pub(crate) fn jwt(&self) -> Jwt {
+        Jwt::new_secp256k1(&self.wallet)
+    }
+}
+
+impl From<bip39::Mnemonic> for Account {
+    fn from(mnemonic: bip39::Mnemonic) -> Self {
+        let wallet = DirectSecp256k1HdWallet::from_mnemonic("n", mnemonic);
+        Self { wallet }
+    }
+}
+
+pub struct Device {
+    keypair: Arc<ed25519::KeyPair>,
+}
+
+impl Device {
+    pub(crate) fn identity_key(&self) -> &ed25519::PublicKey {
+        self.keypair.public_key()
+    }
+
+    pub(crate) fn jwt(&self) -> Jwt {
+        Jwt::new_ecdsa(&self.keypair)
+    }
+}
+
+impl From<Arc<ed25519::KeyPair>> for Device {
+    fn from(keypair: Arc<ed25519::KeyPair>) -> Self {
+        Self { keypair }
+    }
+}
+
+impl From<ed25519::KeyPair> for Device {
+    fn from(keypair: ed25519::KeyPair) -> Self {
+        Self {
+            keypair: Arc::new(keypair),
+        }
+    }
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize)]
+pub struct Gateway {
+    pub identity_key: String,
+    pub location: Location,
+    pub last_probe: Option<Probe>,
+}
+
+impl Gateway {
+    pub fn is_fully_operational_entry(&self) -> bool {
+        self.last_probe
+            .as_ref()
+            .map(|probe| probe.is_fully_operational_entry())
+            .unwrap_or(false)
+    }
+
+    pub fn is_fully_operational_exit(&self) -> bool {
+        self.last_probe
+            .as_ref()
+            .map(|probe| probe.is_fully_operational_exit())
+            .unwrap_or(false)
+    }
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize)]
+pub struct Location {
+    pub two_letter_iso_country_code: String,
+    pub latitude: f64,
+    pub longitude: f64,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct Probe {
+    pub last_updated_utc: String,
+    pub outcome: ProbeOutcome,
+}
+
+impl Probe {
+    pub fn is_fully_operational_entry(&self) -> bool {
+        if !is_recently_updated(&self.last_updated_utc) {
+            return false;
+        }
+        self.outcome.is_fully_operational_entry()
+    }
+
+    pub fn is_fully_operational_exit(&self) -> bool {
+        if !is_recently_updated(&self.last_updated_utc) {
+            return false;
+        }
+        self.outcome.is_fully_operational_exit()
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct ProbeOutcome {
+    pub as_entry: Entry,
+    pub as_exit: Option<Exit>,
+}
+
+impl ProbeOutcome {
+    pub fn is_fully_operational_entry(&self) -> bool {
+        self.as_entry.can_connect && self.as_entry.can_route
+    }
+
+    pub fn is_fully_operational_exit(&self) -> bool {
+        self.as_entry.can_connect
+            && self.as_entry.can_route
+            && self.as_exit.as_ref().map_or(false, |exit| {
+                exit.can_connect
+                    && exit.can_route_ip_v4
+                    && exit.can_route_ip_external_v4
+                    && exit.can_route_ip_v6
+                    && exit.can_route_ip_external_v6
+            })
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct Entry {
+    pub can_connect: bool,
+    pub can_route: bool,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct Exit {
+    pub can_connect: bool,
+    pub can_route_ip_v4: bool,
+    pub can_route_ip_external_v4: bool,
+    pub can_route_ip_v6: bool,
+    pub can_route_ip_external_v6: bool,
+}
+
+fn is_recently_updated(last_updated_utc: &str) -> bool {
+    if let Ok(last_updated) = last_updated_utc.parse::<chrono::DateTime<chrono::Utc>>() {
+        let now = chrono::Utc::now();
+        let duration = now - last_updated;
+        duration.num_minutes() < MAX_PROBE_RESULT_AGE_MINUTES
+    } else {
+        false
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct Country(String);
+
+impl Country {
+    pub fn iso_code(&self) -> &str {
+        &self.0
+    }
+
+    pub fn into_string(self) -> String {
+        self.0
+    }
+}
+
+impl From<String> for Country {
+    fn from(s: String) -> Self {
+        Self(s)
+    }
+}
